# Some performance improvements in bivariate interpolation